Refactor for submissionsService
