+++ src/app/courses/view-courses/courses-view.component.ts
@@ -7,6 +7,7 @@ import { UserService } from '../../shared/user.service';
 import { CoursesService } from '../courses.service';
 import { Subject } from 'rxjs/Subject';
 import { environment } from '../../../environments/environment';
+import { SubmissionsService } from '../../submissions/submissions.service';
 
 @Component({
   templateUrl: './courses-view.component.html',
@@ -43,7 +44,8 @@ export class CoursesViewComponent implements OnInit, OnDestroy {
     private couchService: CouchService,
     private userService: UserService,
     private route: ActivatedRoute,
-    private coursesService: CoursesService
+    private coursesService: CoursesService,
+    private submissionsService: SubmissionsService
   ) { }
 
   ngOnInit() {
@@ -64,12 +66,12 @@ export class CoursesViewComponent implements OnInit, OnDestroy {
   }
 
   goToExam(stepDetail, stepNum) {
-    this.coursesService.openSubmission({
+    this.submissionsService.openSubmission({
       parentId: stepDetail.exam._id + '@' + this.courseDetail._id,
       parent: stepDetail.exam,
       user: this.userService.get().name,
       type: 'exam' });
-    this.coursesService.submissionUpdated$.subscribe(({ submission }) => {
+    this.submissionsService.submissionUpdated$.subscribe(({ submission }) => {
       this.router.navigate([ './step/' + (stepNum + 1) + '/exam',
         (submission.answers.length + 1) ], { relativeTo: this.route });
     });
